refactor(cli): use the canonical dispatch pattern for alerts and license
The alerts, license and license workflow command groups were the last
enterprise commands using an in-file `enum + impl X { execute(&Config) }`
dispatch. Every other group (database, cluster, node, crdb, logs, module,
rbac) declares its enum separately and dispatches through a
`handle_x_command(conn_mgr, ...)` free function backed by a `*_impl.rs`
module.

Convert the three outliers to that shape:

- move `AlertsCommands`, `LicenseCommands` and `LicenseWorkflowCommands`
  into cli/enterprise.rs as `EnterpriseAlertsCommands`,
  `EnterpriseLicenseCommands` and `EnterpriseLicenseWorkflowCommands`
- add alerts_impl.rs, license_impl.rs and license_workflow_impl.rs
  holding the handler functions
- take `&ConnectionManager` instead of building one from `&Config`, so
  the workflow handlers read profiles off `conn_mgr.config`
- drop the cross-file `super::license::calculate_days_remaining` reach
  from the workflow handlers in favour of a `super::license_impl` import

No behavior change: the clap attributes, subcommand names, aliases and
help text all move verbatim, and the handler bodies are unchanged.

Part of #1049.
